HTC-317: removed statusPreference component and from member reg, updated dropdown for multi-select, updated status for multi-select
