+++ client/src/common/forms/Status.js
@@ -6,7 +6,7 @@
  *
  */
 
-import React from 'react';
+import {React,useEffect, useState} from 'react';
 import propTypes from "prop-types";
 import Dropdown from "./Dropdown";
 
@@ -35,15 +35,31 @@ export const statuses = [
 
 
 function Status(props) {
-    const {onChange, dropdownCSS} = props;
+    const {onChange, dropdownCSS, isDropdownMulti} = props;
+    const [selected, setSelected] = useState("");
 
+    const handleInputChange = (e) => {
+        let values = [];
+        for (let val in e) {
+            values.push(e[val].value);
+        }
+        setSelected(values);
+    }
+
+    {isDropdownMulti && useEffect(() => {
+        // this onChange function is the callback from the parent component
+        onChange(selected);
+        console.log(selected)
+        // that can be used to get the value that is inside the dropdown
+    }, [selected]);}
 
     return (
         <div>
             <Dropdown isSearchable={true} placeholder={"Family Status"}
                       options={statuses}
-                      onChange={onChange}
+                      onChange={isDropdownMulti ? handleInputChange : onChange}
                       dropdownCSS={dropdownCSS}
+                      isMulti={isDropdownMulti}
             />
         </div>
     )
@@ -52,7 +68,8 @@ function Status(props) {
 
 Status.propTypes = {
     onChange: propTypes.func,
-    dropdownCSS: propTypes.object
+    dropdownCSS: propTypes.object,
+    isDropdownMulti: propTypes.bool
 };
